Update project.clj, add random-with-seed functions
This commit updates a whole bunch of deps in project.clj (in preparation
for deployment & more features). It also adds a namespace with random
functions that take an instance of `java.util.Random` so that SVGs can
be regenerated from just a seed.

It's nice to be back on this project :)